@@ -2,12 +2,13 @@ package main
 
 import (
 	"database/sql"
-	"github.com/go-gorp/gorp"
-	"github.com/labstack/echo"
-	"gopkg.in/go-playground/validator.v9"
 	"log"
 	"net/http"
 	"time"
+
+	"github.com/go-gorp/gorp"
+	"github.com/go-playground/validator/v10"
+	"github.com/labstack/echo"
 
 	_ "github.com/lib/pq"
 )
@@ -17,9 +18,9 @@ type Validator struct {
 }
 
 type Comment struct {
-	Id int64 `json:"id" db:"id,primarykey,autoincrement"`
-	Name string `json:"name" db:"name,notnull,default:'名無し',size:200"`
-	Text string `json:"text" db:"text,notnull,size:399"`
+	Id      int64     `json:"id" db:"id,primarykey,autoincrement"`
+	Name    string    `json:"name" db:"name,notnull,default:'名無し',size:200"`
+	Text    string    `json:"text" db:"text,notnull,size:399"`
 	Created time.Time `json:"created" db:"created,notnull"`
 	Updated time.Time `json:"updated" db:"updated,notnull"`
 }
@@ -59,7 +60,7 @@ func main() {
 			"SELECT * FROM comments ORDER BY created desc LIMIT 10")
 		if err != nil {
 			c.Logger().Error("Select: ", err)
-			return c.String(http.StatusBadRequest, "Select: " + err.Error())
+			return c.String(http.StatusBadRequest, "Select: "+err.Error())
 		}
 		return c.JSON(http.StatusOK, comments)
 	})
@@ -68,15 +69,15 @@ func main() {
 		var comment Comment
 		if err := c.Bind(&comment); err != nil {
 			c.Logger().Error("Bind: ", err)
-			return c.String(http.StatusBadRequest, "Bind: " + err.Error())
+			return c.String(http.StatusBadRequest, "Bind: "+err.Error())
 		}
 		if err := c.Validate(&comment); err != nil {
 			c.Logger().Error("Validator: ", err)
-			return c.String(http.StatusBadRequest, "Validate: " + err.Error())
+			return c.String(http.StatusBadRequest, "Validate: "+err.Error())
 		}
 		if err := dbmap.Insert(&comment); err != nil {
 			c.Logger().Error("Insert: ", err)
-			return c.String(http.StatusBadRequest, "Insert: " + err.Error())
+			return c.String(http.StatusBadRequest, "Insert: "+err.Error())
 		}
 		c.Logger().Info("ADDED: %v", comment.Id)
 		return c.JSON(http.StatusCreated, "")
